CLI start working without rnc/cli-server-api
When decoupling the community-cli-plugin from the
@react-native-community/cli-server-api (#45311), a middleware stub was
created to allow a runtime stub to be used in this case. This middleware
should be used so as not to break when the optional cli-server-api
dependency isn't present.

Changelog: [General][Fix] - Fix npm react-native start when
cli-server-api isn't installed

Testing: Forced a runtime exception simulating the package not being
dependent and was able to build rn-tester.